% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{DelaunayOnSphere}
\alias{DelaunayOnSphere}
\title{Spherical Delaunay triangulation}
\usage{
DelaunayOnSphere(vertices, radius = 1, center = c(0, 0, 0), iterations = 5)
}
\arguments{
\item{vertices}{vertices, a numeric matrix with three columns}

\item{radius}{radius of the sphere, a positive number; the vertices will
be projected on this sphere}

\item{center}{center of the sphere, a numeric vector of length three; the
vertices will be projected on this sphere}

\item{iterations}{positive integer, the number of iterations used to
construct the meshes of the spherical faces}
}
\value{
A named list with four fields: xxx
}
\description{
Computes a spherical Delaunay triangulation.
}
\details{
xxx
}
\examples{
library(sphereTessellation)
}
\seealso{
\code{\link{plotDelaunayOnSphere}}
}
