% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{DelaunayOnSphere}
\alias{DelaunayOnSphere}
\title{Spherical Delaunay triangulation}
\usage{
DelaunayOnSphere(vertices, radius = 1, center = c(0, 0, 0), iterations = 5L)
}
\arguments{
\item{vertices}{vertices, a numeric matrix with three columns}

\item{radius}{radius of the sphere, a positive number; the vertices will
be projected on this sphere}

\item{center}{center of the sphere, a numeric vector of length three; the
vertices will be projected on this sphere}

\item{iterations}{positive integer, the number of iterations used to
construct the meshes of the spherical faces}
}
\value{
A named list with four fields: xxx
}
\description{
Computes a spherical Delaunay triangulation.
}
\details{
xxx
}
\examples{
library(sphereTessellation)
library(rgl)

if(require(cooltools)) {
set.seed(421L)
vertices <- fibonaccisphere(15L)
del <- DelaunayOnSphere(vertices)
open3d(windowRect = 50 + c(0, 0, 512, 512), zoom = 0.8)
plotDelaunayOnSphere(del)
}

if(require(uniformly)) {
# sample vertices on a hemisphere, so there will be some ghost faces
set.seed(421L)
vertices <- rphong_on_hemisphere(7L)
del <- DelaunayOnSphere(vertices)
# the ghost faces are not plotted
open3d(windowRect = 50 + c(0, 0, 512, 512), zoom = 0.8)
plotDelaunayOnSphere(del)
}
}
\seealso{
\code{\link{plotDelaunayOnSphere}}
}
