% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{DelaunayOnSphere}
\alias{DelaunayOnSphere}
\title{Spherical Delaunay triangulation}
\usage{
DelaunayOnSphere(vertices, radius = 1, center = c(0, 0, 0), iterations = 5)
}
\arguments{
\item{vertices}{vertices, xxx}

\item{radius}{radius of the sphere}

\item{center}{center of the sphere}

\item{iterations}{xxx}
}
\value{
A list with xxx
}
\description{
Computes a spherical Delaunay triangulation.
}
\details{
xxx
}
\examples{
library(sphereTessellation)
}
\seealso{
\code{\link{plotDelaunayOnSphere}}
}
