% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{plotDelaunayOnSphere}
\alias{plotDelaunayOnSphere}
\title{Plot spherical Delaunay triangulation}
\usage{
plotDelaunayOnSphere(del, colors = "random", edges = FALSE, vertices = FALSE)
}
\arguments{
\item{del}{an output of \code{\link{DelaunayOnSphere}}}

\item{colors}{xxx}

\item{edges}{Boolean, whether to plot the edges}

\item{vertices}{Boolean, whether to plot the vertices}
}
\value{
No value is returned.
}
\description{
Plot a spherical Delaunay triangulation.
}
\examples{
library(sphereTessellation)
library(rgl)

vertices <- t(cuboctahedron3d()$vb[-4L, ])
del <- DelaunayOnSphere(vertices, radius = sqrt(2))

open3d(windowRect = 50 + c(0, 0, 512, 512), zoom = 0.8)
plotDelaunayOnSphere(del)
}
