% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plotVoronoiOnSphere}
\alias{plotVoronoiOnSphere}
\title{Plot spherical Voronoï tessellation}
\usage{
plotVoronoiOnSphere(vor, edges = FALSE, sites = FALSE)
}
\arguments{
\item{vor}{an output of \code{\link{VoronoiOnSphere}}}

\item{edges}{Boolean, whether to plot the edges}

\item{sites}{Boolean, whether to plot the Voronoï sites}
}
\value{
No value is returned.
}
\description{
Plot a spherical Voronoï tessellation.
}
\examples{
library(sphereTessellation)
}
