% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plotVoronoiOnSphere}
\alias{plotVoronoiOnSphere}
\title{Plot spherical Voronoï tessellation}
\usage{
plotVoronoiOnSphere(
  vor,
  colors = "palette",
  palette = "Rocket",
  edges = FALSE,
  sites = FALSE
)
}
\arguments{
\item{vor}{an output of \code{\link{VoronoiOnSphere}}}

\item{edges}{Boolean, whether to plot the edges}

\item{sites}{Boolean, whether to plot the Voronoï sites}
}
\value{
No value is returned.
}
\description{
Plot a spherical Voronoï tessellation.
}
\examples{
library(sphereTessellation)
library(rgl)

vertices <- t(cuboctahedron3d()$vb[-4L, ])
vor <- VoronoiOnSphere(vertices)

open3d(windowRect = 50 + c(0, 0, 512, 512), zoom = 0.8)
plotVoronoiOnSphere(vor)
}
