% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{VoronoiOnSphere}
\alias{VoronoiOnSphere}
\title{Spherical Voronoï tessellation}
\usage{
VoronoiOnSphere(vertices, radius = 1, center = c(0, 0, 0), iterations = 5)
}
\arguments{
\item{vertices}{vertices, xxx}

\item{radius}{radius of the sphere}

\item{center}{center of the sphere}

\item{iterations}{xxx}
}
\value{
A list with xxx
}
\description{
Computes a spherical Voronoï tessellation.
}
\details{
xxx
}
\examples{
library(sphereTessellation)
}
\seealso{
\code{\link{plotVoronoiOnSphere}}
}
